9 annotate (#10)
* Clarify that get_terrain_name() shall never return None

* Only assign integer health values to a Troop

* xy is always a tuple of floats

* Some methods are evaluated just for side effects

* Clarify that 'cities' is in fact a list of City

* Annotate Brush

* Annotate generate_terrain()

* Simplify an "int(1)" constant

* Clarify that the draw_info player is in fact a player number (or index)

* Annotate get_terrain() and helpers

* Troop mentions Player, so define it afterwards, to avoid forward-reference issue

* Change the simple_socket "import" slightly, to reduce "socket.socket" stuttering

* Clarify that lsn() actually accepts a number of connections to listen for, and not a vector of connection objects

* Adjust wod_server.py import to avoid 'socket.socket' references

* empty list: annotate that players_in_cities is list[list[Player]]

* Annotate the client draw_info

* Annotate the client player_input

* Keep DeterministicEnvironment2 near its parent class
